refactor. 좌우 반복 이동
- 공통 기능을 따로 빼서 스크립트 작성(MovingMon.cs)
- 좌우 반복 이동이 필요한 객체들이 이를 상속 받아서 사용할 수 있게 수정(Mollusk, Rabbit)
