+++ BadTrip/Assets/Fungus/Integrations/Playmaker/Fungus-PlayMaker.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 81dbcde0f90df4e9ba9ca2794490e57a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
