+++ BadTrip/Assets/Fungus/Integrations/Spine/Fungus-Spine.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fd02c799f3f5c4c83b2fc26c105a3821
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
