+++ BadTrip/Assets/Prefabs/Dream/Rabbit_Jump.prefab
@@ -11,7 +11,7 @@ GameObject:
   - component: {fileID: 6328951325473437784}
   - component: {fileID: 2480163218423443750}
   - component: {fileID: 8333712734895287168}
-  - component: {fileID: 3156943557196326364}
+  - component: {fileID: 889236817178033171}
   m_Layer: 0
   m_Name: Rabbit_Jump
   m_TagString: Untagged
@@ -107,7 +107,7 @@ Animator:
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorStateOnDisable: 0
   m_WriteDefaultValuesOnDisable: 0
---- !u!114 &3156943557196326364
+--- !u!114 &889236817178033171
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -116,9 +116,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 4995786477354264155}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 773b970b6332fc345b82eb38a2f2570e, type: 3}
+  m_Script: {fileID: 11500000, guid: 0cdd61ee612b3874c9982367771d4c61, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   rightMax: 5
   leftMax: -5
-  velocity: 2
+  velocity: 3
