+++ BadTrip/Assets/Scenes/MiniGames/StageOne.unity
@@ -1047,6 +1047,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 2857059022750851396, guid: 57bfaf25cd246f4489e6b7f6ca571e15, type: 3}
+      propertyPath: target
+      value: 
+      objectReference: {fileID: 1898860422}
     - target: {fileID: 5200748158730733864, guid: 57bfaf25cd246f4489e6b7f6ca571e15, type: 3}
       propertyPath: m_Name
       value: SnakeMon (1)
@@ -8324,6 +8328,10 @@ PrefabInstance:
       value: RabbitMon (2)
       objectReference: {fileID: 0}
     - target: {fileID: 327511800121348170, guid: b6d7311886db17c4d92b78e5478df08e, type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 327511800121348170, guid: b6d7311886db17c4d92b78e5478df08e, type: 3}
       propertyPath: m_TagString
       value: Monster
       objectReference: {fileID: 0}
@@ -8431,6 +8439,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 9009426871441980865, guid: b6d7311886db17c4d92b78e5478df08e, type: 3}
+      propertyPath: target
+      value: 
+      objectReference: {fileID: 1898860422}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
@@ -12465,7 +12477,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 2480163218423443750, guid: 0caff4b8b2d60904f9e6afdfcef634da, type: 3}
       propertyPath: m_SortingLayer
-      value: 5
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 2480163218423443750, guid: 0caff4b8b2d60904f9e6afdfcef634da, type: 3}
       propertyPath: m_SortingOrder
@@ -16786,6 +16798,10 @@ PrefabInstance:
       value: RabbitMon (1)
       objectReference: {fileID: 0}
     - target: {fileID: 327511800121348170, guid: b6d7311886db17c4d92b78e5478df08e, type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 327511800121348170, guid: b6d7311886db17c4d92b78e5478df08e, type: 3}
       propertyPath: m_TagString
       value: Monster
       objectReference: {fileID: 0}
@@ -16873,6 +16889,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 9009426871441980865, guid: b6d7311886db17c4d92b78e5478df08e, type: 3}
+      propertyPath: target
+      value: 
+      objectReference: {fileID: 1898860422}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
@@ -17584,7 +17604,7 @@ Transform:
   m_GameObject: {fileID: 1898860421}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 1796.6, y: -244.9, z: 0}
+  m_LocalPosition: {x: 1555.5, y: -241.6, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -17739,7 +17759,7 @@ AudioSource:
   Priority: 128
   DopplerLevel: 1
   MinDistance: 1
-  MaxDistance: 500
+  MaxDistance: 62.23607
   Pan2D: 0
   rolloffMode: 0
   BypassEffects: 0
@@ -22072,6 +22092,10 @@ PrefabInstance:
       value: RabbitMon
       objectReference: {fileID: 0}
     - target: {fileID: 327511800121348170, guid: b6d7311886db17c4d92b78e5478df08e, type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 327511800121348170, guid: b6d7311886db17c4d92b78e5478df08e, type: 3}
       propertyPath: m_TagString
       value: Monster
       objectReference: {fileID: 0}
@@ -22155,6 +22179,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 9009426871441980865, guid: b6d7311886db17c4d92b78e5478df08e, type: 3}
+      propertyPath: target
+      value: 
+      objectReference: {fileID: 1898860422}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
@@ -22334,6 +22362,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 2857059022750851396, guid: 57bfaf25cd246f4489e6b7f6ca571e15, type: 3}
+      propertyPath: target
+      value: 
+      objectReference: {fileID: 1898860422}
     - target: {fileID: 3908631517664616159, guid: 57bfaf25cd246f4489e6b7f6ca571e15, type: 3}
       propertyPath: target
       value: 
@@ -22735,6 +22767,14 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 353673306}
     m_Modifications:
+    - target: {fileID: 889236817178033171, guid: 0caff4b8b2d60904f9e6afdfcef634da, type: 3}
+      propertyPath: leftMax
+      value: -5
+      objectReference: {fileID: 0}
+    - target: {fileID: 889236817178033171, guid: 0caff4b8b2d60904f9e6afdfcef634da, type: 3}
+      propertyPath: rightMax
+      value: 5
+      objectReference: {fileID: 0}
     - target: {fileID: 3156943557196326364, guid: 0caff4b8b2d60904f9e6afdfcef634da, type: 3}
       propertyPath: leftMax
       value: -3
