+++ BadTrip/Assets/Scripts/MiniGames/Dream/Mollusk.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 773b970b6332fc345b82eb38a2f2570e
+guid: ce98cf0b787a5a745a16efcd51f40f20
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
