+++ BadTrip/Assets/Scripts/MiniGames/Dream/MovingMon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 043290ab840d18a448606788c1d0ab57
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
