+++ BadTrip/Assets/Scripts/MiniGames/Dream/Rabbit.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 54bd731f4ffbc3b41b9eea52a5e26c9d
+guid: 0cdd61ee612b3874c9982367771d4c61
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
